wip(components): split 'TimesSelect' from 'AvailabilitySelect'
