+++ components/availability-select/availability-select.module.scss
@@ -1,0 +1,284 @@
+@use 'sass:map';
+@use 'sass:color';
+@use 'styles/config';
+
+.textField {
+  width: 100%;
+}
+
+.menuSurface {
+  display: flex;
+  visibility: hidden;
+  flex-direction: column;
+
+  &:global(.mdc-menu-surface--animating-open),
+  &:global(.mdc-menu-surface--animating-closed),
+  &:global(.mdc-menu-surface--open) {
+    visibility: visible;
+  }
+}
+
+.headers {
+  flex: none;
+  display: flex;
+  overflow: hidden;
+
+  .space {
+    min-width: 40px + 9px;
+    flex: none;
+  }
+
+  .headerWrapper {
+    width: 81px;
+    min-width: 81px;
+    flex: 1 0 auto;
+    overflow: hidden;
+    box-sizing: border-box;
+    display: flex;
+
+    .headerContent {
+      margin: 0;
+      text-align: center;
+      width: 100%;
+
+      .day {
+        line-height: 32px;
+        position: relative;
+        z-index: 2;
+        color: map.get(config.$colors, 'accents-5');
+        font-size: 11px;
+        font-weight: 600;
+        font-family: map.get(config.$fonts, 'sans');
+        letter-spacing: 0.8px;
+        margin-left: 0;
+        margin-top: 8px;
+        text-indent: 0.8px;
+        text-transform: uppercase;
+      }
+
+      .num {
+        position: relative;
+        z-index: 2;
+        font-size: 26px;
+        letter-spacing: -2.6px;
+        text-indent: -2.6px;
+        font-variant: tabular-nums;
+        -moz-font-feature-settings: 'tnum' 1;
+        font-feature-settings: 'tnum' 1;
+        -moz-border-radius: 100%;
+        border-radius: 100%;
+        color: #3c4043;
+        font-weight: 400;
+        font-family: map.get(config.$fonts, 'sans');
+        line-height: 46px;
+        height: 46px;
+        margin-left: auto;
+        margin-right: auto;
+        margin-top: -8px;
+        width: 46px;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+
+        &:hover {
+          background-color: #f1f3f4;
+        }
+      }
+    }
+  }
+
+  .scroller {
+    overflow: scroll;
+    visibility: hidden;
+    flex: none;
+  }
+}
+
+.headerCells {
+  height: 8px;
+  display: flex;
+  flex: none;
+
+  .space {
+    width: 40px + 9px;
+    flex: none;
+  }
+
+  .headerCell {
+    width: 81px;
+    min-width: 81px;
+    flex: 1 0 auto;
+    border-right: #dadce0 1px solid;
+
+    &:last-child {
+      border-right: white 1px solid;
+    }
+  }
+
+  .scroller {
+    overflow: scroll;
+    visibility: hidden;
+    flex: none;
+  }
+}
+
+.gridWrapper {
+  position: relative;
+  flex: 1 1 60%;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+
+  .grid {
+    display: flex;
+    overflow: hidden;
+    align-items: stretch;
+    flex: 1 1 auto;
+
+    &::before {
+      box-shadow: inset 0 1px 1px 0 rgba(0, 0, 0, 0.14),
+        inset 0 2px 1px -1px rgba(0, 0, 0, 0.12);
+      content: '';
+      height: 4px;
+      position: absolute;
+      width: calc(100% - 16px);
+      z-index: 505;
+    }
+
+    &::after {
+      background-image: linear-gradient(to left, white, rgba(255, 255, 255, 0));
+      content: '';
+      height: 2px;
+      position: absolute;
+      right: 16px;
+      width: 80px;
+      z-index: 505;
+    }
+
+    .timesWrapper {
+      overflow-y: scroll;
+      overflow-x: hidden;
+      -ms-scrollbar-width: none;
+      scrollbar-width: none;
+      flex: none;
+      display: flex;
+      align-items: flex-start;
+      min-width: 40px;
+      height: auto;
+
+      &::-webkit-scrollbar {
+        display: none;
+      }
+
+      &::before {
+        background-image: linear-gradient(
+          to right,
+          white,
+          rgba(255, 255, 255, 0)
+        );
+        content: '';
+        height: 2px;
+        position: absolute;
+        width: 80px;
+        z-index: 505;
+      }
+
+      .times {
+        position: relative;
+        background-color: #fff;
+        box-sizing: border-box;
+        margin-left: auto;
+
+        .timeWrapper {
+          position: relative;
+          height: 48px;
+          padding-right: 8px;
+          text-align: right;
+
+          .timeLabel {
+            display: block;
+            position: relative;
+            top: -6px;
+            color: map.get(config.$colors, 'accents-5');
+            font-size: 10px;
+            font-family: map.get(config.$fonts, 'sans');
+          }
+
+          &:first-child .timeLabel {
+            display: none;
+          }
+        }
+      }
+    }
+
+    .rowsWrapper {
+      overflow-y: scroll;
+      overflow-x: auto;
+      display: flex;
+      align-items: flex-start;
+      flex: 1 1 auto;
+
+      .rows {
+        flex: none;
+        display: flex;
+        min-width: 100%;
+        overflow: hidden;
+        position: relative;
+
+        .lines {
+          border-top: #dadce0 1px solid;
+
+          .line {
+            height: 48px;
+
+            &::after {
+              content: '';
+              border-bottom: #dadce0 1px solid;
+              position: absolute;
+              width: 100%;
+              margin-top: -1px;
+              z-index: 3;
+              pointer-events: none;
+            }
+
+            &:first-child::after {
+              content: none;
+            }
+          }
+        }
+
+        .space {
+          width: 8px;
+          border-right: #dadce0 1px solid;
+        }
+
+        .cells {
+          display: flex;
+          flex: 1 1 auto;
+          box-sizing: border-box;
+          position: relative;
+          $dark-strip: color.adjust(#dadce0, $lightness: 5%);
+          $light-strip: color.adjust($dark-strip, $lightness: 5%);
+          background: repeating-linear-gradient(
+            45deg,
+            $light-strip,
+            $light-strip 10px,
+            $dark-strip 10px,
+            $dark-strip 20px
+          );
+
+          .cell {
+            width: 81px;
+            min-width: 81px;
+            flex: 1 0 auto;
+            border-right: #dadce0 1px solid;
+
+            &:last-child {
+              border-right: white 1px solid;
+            }
+          }
+        }
+      }
+    }
+  }
+}
