+++ components/availability-select/global.d.ts
