+++ components/availability-select/utils.ts
