+++ components/times-select/index.tsx
@@ -1,24 +1,9 @@
 import { MenuSurface, MenuSurfaceAnchor } from '@rmwc/menu';
-import {
-  MouseEvent,
-  useCallback,
-  useLayoutEffect,
-  useRef,
-  useState,
-} from 'react';
-import { ScrollSync, ScrollSyncPane } from 'react-scroll-sync';
 import { TextField, TextFieldHTMLProps, TextFieldProps } from '@rmwc/textfield';
-import { ResizeObserver as polyfill } from '@juggle/resize-observer';
-import { v4 as uuid } from 'uuid';
-import useMeasure from 'react-use-measure';
-import useTranslation from 'next-translate/useTranslation';
+import { useCallback, useLayoutEffect, useRef, useState } from 'react';
 
-import { TimeUtils } from 'lib/utils';
-import { Availability, DayAlias, TCallback, Timeslot } from 'lib/model';
+import { Availability, Callback } from 'lib/model';
 
-import OptionRnd from './option-rnd';
-import TimeslotRnd from './timeslot-rnd';
-import { getTimeslot } from './utils';
 import styles from './times-select.module.scss';
 
 type OverridenProps =
@@ -30,7 +15,7 @@ type OverridenProps =
   | 'className';
 interface Props {
   value: Availability;
-  onChange: TCallback<Availability>;
+  onChange: Callback<Availability>;
   options?: Availability;
   renderToPortal?: boolean;
   focused?: boolean;
@@ -55,7 +40,7 @@ export type TimesSelectProps = Omit<
 export default function TimesSelect({
   value,
   onChange,
-  options = Availability.full(),
+  options,
   renderToPortal,
   focused,
   onFocused,
@@ -63,28 +48,16 @@ export default function TimesSelect({
   className,
   ...textFieldProps
 }: TimesSelectProps): JSX.Element {
-  const { lang: locale } = useTranslation();
-
-  const rowsRef = useRef<HTMLDivElement | null>(null);
   const inputRef = useRef<HTMLInputElement | null>(null);
-  const timeoutId = useRef<ReturnType<typeof setTimeout>>();
-
-  const [cellsRef, { x, y }] = useMeasure({ polyfill, scroll: true });
-  const [scrolled, setScrolled] = useState<boolean>(false);
-  const [menuOpen, setMenuOpen] = useState<boolean>(false);
-
   useLayoutEffect(() => {
     if (focused && inputRef.current) inputRef.current.focus();
   }, [focused]);
 
-  useLayoutEffect(() => {
-    // Scroll to 8:30am by default (assumes 48px per hour).
-    if (rowsRef.current && !scrolled) rowsRef.current.scrollTop = 48 * 8 + 24;
-  }, [scrolled, menuOpen]);
-
   // We use `setTimeout` and `clearTimeout` to wait a "tick" on a blur event
   // before toggling which ensures the user hasn't re-opened the menu.
   // @see {@link https:bit.ly/2x9eM27}
+  const [menuOpen, setMenuOpen] = useState<boolean>(false);
+  const timeoutId = useRef<ReturnType<typeof setTimeout>>();
   const openMenu = useCallback(() => {
     if (timeoutId.current) {
       clearTimeout(timeoutId.current);
@@ -96,17 +69,6 @@ export default function TimesSelect({
     timeoutId.current = setTimeout(() => setMenuOpen(false), 0);
   }, []);
 
-  // Create a new `TimeslotRND` closest to the user's click position. Assumes
-  // each column is 82px wide and every hour is 48px tall (i.e. 12px = 15min).
-  const onClick = useCallback(
-    (event: MouseEvent) => {
-      const position = { x: event.pageX - x, y: event.pageY - y };
-      onChange(new Availability(...value, getTimeslot(48, position)));
-    },
-    [x, y, onChange, value]
-  );
-  const onScroll = useCallback(() => setScrolled(true), []);
-
   return (
     <MenuSurfaceAnchor className={className}>
       <MenuSurface
@@ -115,105 +77,9 @@ export default function TimesSelect({
         onFocus={openMenu}
         onBlur={closeMenu}
         anchorCorner='bottomStart'
-        className={styles.menuSurface}
         renderToPortal={renderToPortal ? '#portal' : false}
       >
-        <div className={styles.headers}>
-          <div className={styles.space} />
-          {Array(7)
-            .fill(null)
-            .map((_: null, weekday: number) => (
-              <div key={weekday} className={styles.headerWrapper}>
-                <h2 className={styles.headerContent}>
-                  <div className={styles.day}>
-                    {TimeUtils.getNextDateWithDay(weekday as DayAlias)
-                      .toLocaleString(locale, { weekday: 'long' })
-                      .substr(0, 3)}
-                  </div>
-                </h2>
-              </div>
-            ))}
-          <div className={styles.scroller} />
-        </div>
-        <div className={styles.headerCells}>
-          <div className={styles.space} />
-          {Array(7)
-            .fill(null)
-            .map(() => (
-              <div key={uuid()} className={styles.headerCell} />
-            ))}
-          <div className={styles.scroller} />
-        </div>
-        <div className={styles.gridWrapper}>
-          <ScrollSync>
-            <div className={styles.grid}>
-              <ScrollSyncPane>
-                <div className={styles.timesWrapper}>
-                  <div className={styles.times}>
-                    {Array(24)
-                      .fill(null)
-                      .map((_: null, hour: number) => (
-                        <div key={hour} className={styles.timeWrapper}>
-                          <span className={styles.timeLabel}>
-                            {TimeUtils.getDateWithTime(
-                              hour
-                            ).toLocaleString(locale, { hour: '2-digit' })}
-                          </span>
-                        </div>
-                      ))}
-                  </div>
-                </div>
-              </ScrollSyncPane>
-              <ScrollSyncPane>
-                <div
-                  className={styles.rowsWrapper}
-                  ref={rowsRef}
-                  onScroll={onScroll}
-                >
-                  <div className={styles.rows}>
-                    <div className={styles.lines}>
-                      {Array(24)
-                        .fill(null)
-                        .map(() => (
-                          <div key={uuid()} className={styles.line} />
-                        ))}
-                    </div>
-                    <div className={styles.space} />
-                    <div
-                      className={styles.cells}
-                      onClick={onClick}
-                      ref={cellsRef}
-                    >
-                      {options.map((timeslot: Timeslot) => (
-                        <OptionRnd key={uuid()} value={timeslot} />
-                      ))}
-                      {value.map((timeslot: Timeslot, idx: number) => (
-                        <TimeslotRnd
-                          key={uuid()}
-                          value={timeslot}
-                          onChange={(updated: Timeslot) => {
-                            onChange(
-                              new Availability(
-                                ...value.slice(0, idx),
-                                updated,
-                                ...value.slice(idx + 1)
-                              )
-                            );
-                          }}
-                        />
-                      ))}
-                      {Array(7)
-                        .fill(null)
-                        .map(() => (
-                          <div key={uuid()} className={styles.cell} />
-                        ))}
-                    </div>
-                  </div>
-                </div>
-              </ScrollSyncPane>
-            </div>
-          </ScrollSync>
-        </div>
+        <div className={styles.wrapper} />
       </MenuSurface>
       <TextField
         {...textFieldProps}
@@ -221,7 +87,7 @@ export default function TimesSelect({
         textarea={false}
         inputRef={inputRef}
         value={value.toString()}
-        className={styles.textField}
+        className={styles.field}
         onFocus={() => {
           if (onFocused) onFocused();
           openMenu();
